feat(producer): refactor encap and packet decoding
